requirements.txt
# Video Frame Search System Requirements
# Python 3.8+ required

# Core dependencies
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3
pandas==2.1.3
tqdm==4.66.1
python-dotenv==1.0.0

# Deep Learning frameworks
torch==2.1.0
torchvision==0.16.0
transformers==4.35.2
sentence-transformers==2.2.2
# For BGE model support
FlagEmbedding>=1.2.0

# Vector Database
pinecone>=3.0.0

# Optional but recommended
scipy==1.11.4
scikit-learn==1.3.2
matplotlib==3.8.1

# Development tools (optional)
pytest==7.4.3
black==23.11.0
flake8==6.1.0
ipython==8.17.2
jupyter==1.0.0
notebook==7.0.6